naming: propagate renaming to `automation-tools`
